+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Fixed handling of errors in recursive QCJSON creator operations.
 - Support for xtb optimizations, single-point energies, and molecular dynamics. (xtb)
 - Support for additional functionals. (ORCA)